ai-module/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.8
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
scipy==1.11.3
Pillow==10.1.0